LLP Messages / Checksums / Debugging

+++ src/LLP/include/node.h
@@ -38,7 +38,7 @@ namespace LLP
 		
 		
 		/* Node Identifier */
-		unsigned int nSessionID;
+		uint64 nSessionID;
 		
 		
 		/* Basic Stats. */
